nlp_service/requirements.txt
spacy==2.3.2
joblib
grpcio-tools